+++ aws-lambda/lib/extended-nodejs-function.ts
@@ -3,12 +3,7 @@ import {FunctionAlarms, FunctionAlarmsOptions} from './function-alarms';
 import {FunctionDeployment} from './function-deployment';
 import {DeployedFunctionOptions} from './extended-function';
 import {RetentionDays} from 'aws-cdk-lib/aws-logs';
-import {
-  Architecture,
-  LoggingFormat,
-  Runtime,
-  SystemLogLevel,
-} from 'aws-cdk-lib/aws-lambda';
+import {Architecture, LoggingFormat, Runtime} from 'aws-cdk-lib/aws-lambda';
 import {Duration} from 'aws-cdk-lib';
 import * as fs from 'fs';
 import * as path from 'path';
@@ -60,11 +55,10 @@ export class ExtendedNodejsFunction extends NodejsFunction {
       architecture: Architecture.ARM_64, // change default from X86_64
       memorySize: 768, // change default from 128
       timeout: Duration.seconds(30), // change default from 3
-      runtime: Runtime.NODEJS_20_X, // change default from NODEJS_20_X
+      runtime: Runtime.NODEJS_22_X,
       depsLockFilePath: ExtendedNodejsFunction.findDepsLockFile(props.entry),
       ...props,
       loggingFormat: props.loggingFormat ?? LoggingFormat.JSON,
-      systemLogLevelV2: SystemLogLevel.INFO,
       environment: {
         NODE_OPTIONS: '--enable-source-maps',
         ...props.environment,