+++ CHANGELOG.md
@@ -5,11 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v18.0.0
+## v1.18.0
+
+### Added
+
+-   Latency based Cname records to aws-route53
+-   Added ExtendedBucket and refactored WebsiteBucket and CloudFrontBucketV2 to use it
 
 ### Changed
 
+-   Update ExtendedNodejsFunction and CdkPipeline to use Node 22 runtime as the default
+-   Modified StandardTable.addGlobalSecondaryIndex to allow for more flexible index definitions
+-   Simplified deploy methods on WebsiteBucket and CloudFrontBucketV2
 -   Updated dependencies to newer versions
+
+## v1.17.2
+
+### Added
+
+-   Added AWS::OpenSearchServerless::Collection to EXCLUDED_RESOURCES for tagging.
+
+### Fixed
+
+-   Fixed issue where excludeResourceTypes was being overridden on StandardTags
+
+## v1.17.1
+
+### Changed
+
+-   Updated dependencies
 
 ## v1.17.0
 
@@ -17,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added exportAndOutputParameter to ExtendedStack to allow for exporting and outputting parameters in the stack with one call
 -   Default logging for all extended lambda functions is now JSON instead of TEXT
+-   Updated dependencies to newer versions
 
 ### Removed